# Optimize Buffer.utf8ToString use String ctor