Make `Buffer.utf8ToString` simply wrap the String decode constructor, this is to take advantage of the compact string optimization (which we've observed to provide faster performance for many of our clients) and in general for code maintainability.

This is the case since https://bugs.openjdk.org/browse/JDK-8259842 which removed the TLS overhead that I've seen you've also encountered based on your comments.

Built and ran tests, not all tests pass but I'm getting the same errors without the change as well.

```plain
Tests run: 309, Failures: 11, Errors: 6, Skipped: 4, Time elapsed: 105.231 sec <<< FAILURE! - in com.aerospike.test.SuiteAll
txnDeleteAbort(com.aerospike.test.sync.basic.TestTxn)  Time elapsed: 0.016 sec  <<< FAILURE!
java.lang.AssertionError: Failed to get: namespace=test set=test key=mrtkey7
        at org.junit.Assert.fail(Assert.java:89)
        at com.aerospike.test.sync.TestSync.assertRecordFound(TestSync.java:53)
        at com.aerospike.test.sync.TestSync.assertBinEqual(TestSync.java:69)
        at com.aerospike.test.sync.basic.TestTxn.txnDeleteAbort(TestTxn.java:222)

txnBatch(com.aerospike.test.sync.basic.TestTxn)  Time elapsed: 0.006 sec  <<< ERROR!
com.aerospike.client.AerospikeException$BatchRecordArray: Error -16: Batch failed
        at com.aerospike.client.command.MultiCommand.parseGroup(MultiCommand.java:236)
        at com.aerospike.client.command.MultiCommand.parseResult(MultiCommand.java:219)
        at com.aerospike.client.command.SyncCommand.executeCommand(SyncCommand.java:111)
        at com.aerospike.client.command.SyncCommand.execute(SyncCommand.java:66)
        at com.aerospike.client.command.BatchExecutor.execute(BatchExecutor.java:35)
        at com.aerospike.client.AerospikeClient.operate(AerospikeClient.java:3032)
        at com.aerospike.test.sync.basic.TestTxn.txnBatch(TestTxn.java:383)

txnUDFAbort(com.aerospike.test.sync.basic.TestTxn)  Time elapsed: 0.005 sec  <<< FAILURE!
java.lang.AssertionError: Data mismatch: Expected val1. Received val2
        at org.junit.Assert.fail(Assert.java:89)
        at com.aerospike.test.sync.TestSync.assertBinEqual(TestSync.java:74)
        at com.aerospike.test.sync.basic.TestTxn.txnUDFAbort(TestTxn.java:358)

txnWriteAbort(com.aerospike.test.sync.basic.TestTxn)  Time elapsed: 0.004 sec  <<< FAILURE!
java.lang.AssertionError: Data mismatch: Expected val1. Received val2
        at org.junit.Assert.fail(Assert.java:89)
        at com.aerospike.test.sync.TestSync.assertBinEqual(TestSync.java:74)
        at com.aerospike.test.sync.basic.TestTxn.txnWriteAbort(TestTxn.java:184)

txnWriteBlock(com.aerospike.test.sync.basic.TestTxn)  Time elapsed: 0.004 sec  <<< ERROR!
com.aerospike.client.AerospikeException: Error -1: Unexpected success
        at com.aerospike.test.sync.basic.TestTxn.txnWriteBlock(TestTxn.java:132)

txnWriteConflict(com.aerospike.test.sync.basic.TestTxn)  Time elapsed: 0.004 sec  <<< FAILURE!
java.lang.AssertionError: Data mismatch: Expected val1. Received val2
        at org.junit.Assert.fail(Assert.java:89)
        at com.aerospike.test.sync.TestSync.assertBinEqual(TestSync.java:74)
        at com.aerospike.test.sync.basic.TestTxn.txnWriteConflict(TestTxn.java:114)

txnBatchAbort(com.aerospike.test.sync.basic.TestTxn)  Time elapsed: 0.006 sec  <<< ERROR!
com.aerospike.client.AerospikeException$BatchRecordArray: Error -16: Batch failed
        at com.aerospike.client.command.MultiCommand.parseGroup(MultiCommand.java:236)
        at com.aerospike.client.command.MultiCommand.parseResult(MultiCommand.java:219)
        at com.aerospike.client.command.SyncCommand.executeCommand(SyncCommand.java:111)
        at com.aerospike.client.command.SyncCommand.execute(SyncCommand.java:66)
        at com.aerospike.client.command.BatchExecutor.execute(BatchExecutor.java:35)
        at com.aerospike.client.AerospikeClient.operate(AerospikeClient.java:3032)
        at com.aerospike.test.sync.basic.TestTxn.txnBatchAbort(TestTxn.java:431)

txnOperateWriteAbort(com.aerospike.test.sync.basic.TestTxn)  Time elapsed: 0.003 sec  <<< FAILURE!
java.lang.AssertionError: Data mismatch: Expected val1. Received val2
        at org.junit.Assert.fail(Assert.java:89)
        at com.aerospike.test.sync.TestSync.assertBinEqual(TestSync.java:74)
        at com.aerospike.test.sync.basic.TestTxn.txnOperateWriteAbort(TestTxn.java:322)

txnWriteRead(com.aerospike.test.sync.basic.TestTxn)  Time elapsed: 0.001 sec  <<< FAILURE!
java.lang.AssertionError: Data mismatch: Expected val1. Received val2
        at org.junit.Assert.fail(Assert.java:89)
        at com.aerospike.test.sync.TestSync.assertBinEqual(TestSync.java:74)
        at com.aerospike.test.sync.basic.TestTxn.txnWriteRead(TestTxn.java:156)

asyncTxnBatchAbort(com.aerospike.test.async.TestAsyncTxn)  Time elapsed: 0.012 sec  <<< ERROR!
com.aerospike.client.AerospikeException: Error 4,1,0,30000,1000,0,BB9020011AC4202 127.0.0.1 3000: Parameter error
        at com.aerospike.client.async.AsyncMultiCommand.parseResult(AsyncMultiCommand.java:78)
        at com.aerospike.client.async.AsyncCommand.parseCommandResult(AsyncCommand.java:173)
        at com.aerospike.client.async.NioCommand.parseGroupBody(NioCommand.java:651)
        at com.aerospike.client.async.NioCommand.readMultiBody(NioCommand.java:590)
        at com.aerospike.client.async.NioCommand.readMultiHeader(NioCommand.java:582)
        at com.aerospike.client.async.NioCommand.read(NioCommand.java:494)
        at com.aerospike.client.async.NioCommand.processEvent(NioCommand.java:325)
        at com.aerospike.client.async.NioEventLoop.runCommands(NioEventLoop.java:223)
        at com.aerospike.client.async.NioEventLoop.run(NioEventLoop.java:181)
        at java.base/java.lang.Thread.run(Thread.java:1583)

asyncTxnWriteRead(com.aerospike.test.async.TestAsyncTxn)  Time elapsed: 0.003 sec  <<< FAILURE!
java.lang.AssertionError: java.lang.Exception: Data mismatch: Expected val1. Received val2
        at com.aerospike.test.async.TestAsync.assertBinEqual(TestAsync.java:78)
        at com.aerospike.test.async.TestAsyncTxn$GetExpect$1.onSuccess(TestAsyncTxn.java:483)
        at com.aerospike.client.async.AsyncRead.onSuccess(AsyncRead.java:80)
        at com.aerospike.client.async.NioCommand.finish(NioCommand.java:822)
        at com.aerospike.client.async.NioCommand.readSingleBody(NioCommand.java:566)
        at com.aerospike.client.async.NioCommand.readSingleHeader(NioCommand.java:554)
        at com.aerospike.client.async.NioCommand.read(NioCommand.java:491)
        at com.aerospike.client.async.NioCommand.processEvent(NioCommand.java:325)
        at com.aerospike.client.async.NioEventLoop.runCommands(NioEventLoop.java:223)
        at com.aerospike.client.async.NioEventLoop.run(NioEventLoop.java:181)
        at java.base/java.lang.Thread.run(Thread.java:1583)

        at org.junit.Assert.fail(Assert.java:89)
        at com.aerospike.test.async.AsyncMonitor.waitTillComplete(AsyncMonitor.java:49)
        at com.aerospike.test.async.TestAsync.waitTillComplete(TestAsync.java:210)
        at com.aerospike.test.async.TestAsyncTxn.execute(TestAsyncTxn.java:341)
        at com.aerospike.test.async.TestAsyncTxn.asyncTxnWriteRead(TestAsyncTxn.java:123)

asyncTxnBatch(com.aerospike.test.async.TestAsyncTxn)  Time elapsed: 0.007 sec  <<< ERROR!
com.aerospike.client.AerospikeException: Error 4,1,0,30000,1000,0,BB9020011AC4202 127.0.0.1 3000: Parameter error
        at com.aerospike.client.async.AsyncMultiCommand.parseResult(AsyncMultiCommand.java:78)
        at com.aerospike.client.async.AsyncCommand.parseCommandResult(AsyncCommand.java:173)
        at com.aerospike.client.async.NioCommand.parseGroupBody(NioCommand.java:651)
        at com.aerospike.client.async.NioCommand.readMultiBody(NioCommand.java:590)
        at com.aerospike.client.async.NioCommand.readMultiHeader(NioCommand.java:582)
        at com.aerospike.client.async.NioCommand.read(NioCommand.java:494)
        at com.aerospike.client.async.NioCommand.processEvent(NioCommand.java:325)
        at com.aerospike.client.async.NioEventLoop.runCommands(NioEventLoop.java:223)
        at com.aerospike.client.async.NioEventLoop.run(NioEventLoop.java:181)
        at java.base/java.lang.Thread.run(Thread.java:1583)

asyncTxnUDFAbort(com.aerospike.test.async.TestAsyncTxn)  Time elapsed: 0.004 sec  <<< FAILURE!
java.lang.AssertionError: java.lang.Exception: Data mismatch: Expected val1. Received val2
        at com.aerospike.test.async.TestAsync.assertBinEqual(TestAsync.java:78)
        at com.aerospike.test.async.TestAsyncTxn$GetExpect$1.onSuccess(TestAsyncTxn.java:483)
        at com.aerospike.client.async.AsyncRead.onSuccess(AsyncRead.java:80)
        at com.aerospike.client.async.NioCommand.finish(NioCommand.java:822)
        at com.aerospike.client.async.NioCommand.readSingleBody(NioCommand.java:566)
        at com.aerospike.client.async.NioCommand.readSingleHeader(NioCommand.java:554)
        at com.aerospike.client.async.NioCommand.read(NioCommand.java:491)
        at com.aerospike.client.async.NioCommand.processEvent(NioCommand.java:325)
        at com.aerospike.client.async.NioEventLoop.runCommands(NioEventLoop.java:223)
        at com.aerospike.client.async.NioEventLoop.run(NioEventLoop.java:181)
        at java.base/java.lang.Thread.run(Thread.java:1583)

        at org.junit.Assert.fail(Assert.java:89)
        at com.aerospike.test.async.AsyncMonitor.waitTillComplete(AsyncMonitor.java:49)
        at com.aerospike.test.async.TestAsync.waitTillComplete(TestAsync.java:210)
        at com.aerospike.test.async.TestAsyncTxn.execute(TestAsyncTxn.java:341)
        at com.aerospike.test.async.TestAsyncTxn.asyncTxnUDFAbort(TestAsyncTxn.java:287)

asyncTxnDeleteAbort(com.aerospike.test.async.TestAsyncTxn)  Time elapsed: 0.003 sec  <<< FAILURE!
java.lang.AssertionError: java.lang.Exception: Failed to get: namespace=test set=test key=asyncTxnDeleteAbort
        at com.aerospike.test.async.TestAsync.assertRecordFound(TestAsync.java:100)
        at com.aerospike.test.async.TestAsync.assertBinEqual(TestAsync.java:71)
        at com.aerospike.test.async.TestAsyncTxn$GetExpect$1.onSuccess(TestAsyncTxn.java:483)
        at com.aerospike.client.async.AsyncRead.onSuccess(AsyncRead.java:80)
        at com.aerospike.client.async.NioCommand.finish(NioCommand.java:822)
        at com.aerospike.client.async.NioCommand.readSingleBody(NioCommand.java:566)
        at com.aerospike.client.async.NioCommand.readSingleHeader(NioCommand.java:554)
        at com.aerospike.client.async.NioCommand.read(NioCommand.java:491)
        at com.aerospike.client.async.NioCommand.processEvent(NioCommand.java:325)
        at com.aerospike.client.async.NioEventLoop.runCommands(NioEventLoop.java:223)
        at com.aerospike.client.async.NioEventLoop.run(NioEventLoop.java:181)
        at java.base/java.lang.Thread.run(Thread.java:1583)

        at org.junit.Assert.fail(Assert.java:89)
        at com.aerospike.test.async.AsyncMonitor.waitTillComplete(AsyncMonitor.java:49)
        at com.aerospike.test.async.TestAsync.waitTillComplete(TestAsync.java:210)
        at com.aerospike.test.async.TestAsyncTxn.execute(TestAsyncTxn.java:341)
        at com.aerospike.test.async.TestAsyncTxn.asyncTxnDeleteAbort(TestAsyncTxn.java:169)

asyncTxnOperateWriteAbort(com.aerospike.test.async.TestAsyncTxn)  Time elapsed: 0.002 sec  <<< FAILURE!
java.lang.AssertionError: java.lang.Exception: Data mismatch: Expected val1. Received val2
        at com.aerospike.test.async.TestAsync.assertBinEqual(TestAsync.java:78)
        at com.aerospike.test.async.TestAsyncTxn$GetExpect$1.onSuccess(TestAsyncTxn.java:483)
        at com.aerospike.client.async.AsyncRead.onSuccess(AsyncRead.java:80)
        at com.aerospike.client.async.NioCommand.finish(NioCommand.java:822)
        at com.aerospike.client.async.NioCommand.readSingleBody(NioCommand.java:566)
        at com.aerospike.client.async.NioCommand.readSingleHeader(NioCommand.java:554)
        at com.aerospike.client.async.NioCommand.read(NioCommand.java:491)
        at com.aerospike.client.async.NioCommand.processEvent(NioCommand.java:325)
        at com.aerospike.client.async.NioEventLoop.runCommands(NioEventLoop.java:223)
        at com.aerospike.client.async.NioEventLoop.run(NioEventLoop.java:181)
        at java.base/java.lang.Thread.run(Thread.java:1583)

        at org.junit.Assert.fail(Assert.java:89)
        at com.aerospike.test.async.AsyncMonitor.waitTillComplete(AsyncMonitor.java:49)
        at com.aerospike.test.async.TestAsync.waitTillComplete(TestAsync.java:210)
        at com.aerospike.test.async.TestAsyncTxn.execute(TestAsyncTxn.java:341)
        at com.aerospike.test.async.TestAsyncTxn.asyncTxnOperateWriteAbort(TestAsyncTxn.java:255)

asyncTxnWriteAbort(com.aerospike.test.async.TestAsyncTxn)  Time elapsed: 0.004 sec  <<< FAILURE!
java.lang.AssertionError: java.lang.Exception: Data mismatch: Expected val1. Received val2
        at com.aerospike.test.async.TestAsync.assertBinEqual(TestAsync.java:78)
        at com.aerospike.test.async.TestAsyncTxn$GetExpect$1.onSuccess(TestAsyncTxn.java:483)
        at com.aerospike.client.async.AsyncRead.onSuccess(AsyncRead.java:80)
        at com.aerospike.client.async.NioCommand.finish(NioCommand.java:822)
        at com.aerospike.client.async.NioCommand.readSingleBody(NioCommand.java:566)
        at com.aerospike.client.async.NioCommand.readSingleHeader(NioCommand.java:554)
        at com.aerospike.client.async.NioCommand.read(NioCommand.java:491)
        at com.aerospike.client.async.NioCommand.processEvent(NioCommand.java:325)
        at com.aerospike.client.async.NioEventLoop.runCommands(NioEventLoop.java:223)
        at com.aerospike.client.async.NioEventLoop.run(NioEventLoop.java:181)
        at java.base/java.lang.Thread.run(Thread.java:1583)

        at org.junit.Assert.fail(Assert.java:89)
        at com.aerospike.test.async.AsyncMonitor.waitTillComplete(AsyncMonitor.java:49)
        at com.aerospike.test.async.TestAsync.waitTillComplete(TestAsync.java:210)
        at com.aerospike.test.async.TestAsyncTxn.execute(TestAsyncTxn.java:341)
        at com.aerospike.test.async.TestAsyncTxn.asyncTxnWriteAbort(TestAsyncTxn.java:139)

asyncTxnWriteBlock(com.aerospike.test.async.TestAsyncTxn)  Time elapsed: 0.002 sec  <<< ERROR!
com.aerospike.client.AerospikeException: Error -1: Unexpected success
        at com.aerospike.test.async.TestAsyncTxn.asyncTxnWriteBlock(TestAsyncTxn.java:101)


Results :

Failed tests: 
  TestAsyncTxn.asyncTxnDeleteAbort:169->execute:341->TestAsync.waitTillComplete:210 java.lang.Exception: Failed to get: namespace=test set=test key=asyncTxnDeleteAbort
        at com.aerospike.test.async.TestAsync.assertRecordFound(TestAsync.java:100)
        at com.aerospike.test.async.TestAsync.assertBinEqual(TestAsync.java:71)
        at com.aerospike.test.async.TestAsyncTxn$GetExpect$1.onSuccess(TestAsyncTxn.java:483)
        at com.aerospike.client.async.AsyncRead.onSuccess(AsyncRead.java:80)
        at com.aerospike.client.async.NioCommand.finish(NioCommand.java:822)
        at com.aerospike.client.async.NioCommand.readSingleBody(NioCommand.java:566)
        at com.aerospike.client.async.NioCommand.readSingleHeader(NioCommand.java:554)
        at com.aerospike.client.async.NioCommand.read(NioCommand.java:491)
        at com.aerospike.client.async.NioCommand.processEvent(NioCommand.java:325)
        at com.aerospike.client.async.NioEventLoop.runCommands(NioEventLoop.java:223)
        at com.aerospike.client.async.NioEventLoop.run(NioEventLoop.java:181)
        at java.base/java.lang.Thread.run(Thread.java:1583)

  TestAsyncTxn.asyncTxnOperateWriteAbort:255->execute:341->TestAsync.waitTillComplete:210 java.lang.Exception: Data mismatch: Expected val1. Received val2
        at com.aerospike.test.async.TestAsync.assertBinEqual(TestAsync.java:78)
        at com.aerospike.test.async.TestAsyncTxn$GetExpect$1.onSuccess(TestAsyncTxn.java:483)
        at com.aerospike.client.async.AsyncRead.onSuccess(AsyncRead.java:80)
        at com.aerospike.client.async.NioCommand.finish(NioCommand.java:822)
        at com.aerospike.client.async.NioCommand.readSingleBody(NioCommand.java:566)
        at com.aerospike.client.async.NioCommand.readSingleHeader(NioCommand.java:554)
        at com.aerospike.client.async.NioCommand.read(NioCommand.java:491)
        at com.aerospike.client.async.NioCommand.processEvent(NioCommand.java:325)
        at com.aerospike.client.async.NioEventLoop.runCommands(NioEventLoop.java:223)
        at com.aerospike.client.async.NioEventLoop.run(NioEventLoop.java:181)
        at java.base/java.lang.Thread.run(Thread.java:1583)

  TestAsyncTxn.asyncTxnUDFAbort:287->execute:341->TestAsync.waitTillComplete:210 java.lang.Exception: Data mismatch: Expected val1. Received val2
        at com.aerospike.test.async.TestAsync.assertBinEqual(TestAsync.java:78)
        at com.aerospike.test.async.TestAsyncTxn$GetExpect$1.onSuccess(TestAsyncTxn.java:483)
        at com.aerospike.client.async.AsyncRead.onSuccess(AsyncRead.java:80)
        at com.aerospike.client.async.NioCommand.finish(NioCommand.java:822)
        at com.aerospike.client.async.NioCommand.readSingleBody(NioCommand.java:566)
        at com.aerospike.client.async.NioCommand.readSingleHeader(NioCommand.java:554)
        at com.aerospike.client.async.NioCommand.read(NioCommand.java:491)
        at com.aerospike.client.async.NioCommand.processEvent(NioCommand.java:325)
        at com.aerospike.client.async.NioEventLoop.runCommands(NioEventLoop.java:223)
        at com.aerospike.client.async.NioEventLoop.run(NioEventLoop.java:181)
        at java.base/java.lang.Thread.run(Thread.java:1583)

  TestAsyncTxn.asyncTxnWriteAbort:139->execute:341->TestAsync.waitTillComplete:210 java.lang.Exception: Data mismatch: Expected val1. Received val2
        at com.aerospike.test.async.TestAsync.assertBinEqual(TestAsync.java:78)
        at com.aerospike.test.async.TestAsyncTxn$GetExpect$1.onSuccess(TestAsyncTxn.java:483)
        at com.aerospike.client.async.AsyncRead.onSuccess(AsyncRead.java:80)
        at com.aerospike.client.async.NioCommand.finish(NioCommand.java:822)
        at com.aerospike.client.async.NioCommand.readSingleBody(NioCommand.java:566)
        at com.aerospike.client.async.NioCommand.readSingleHeader(NioCommand.java:554)
        at com.aerospike.client.async.NioCommand.read(NioCommand.java:491)
        at com.aerospike.client.async.NioCommand.processEvent(NioCommand.java:325)
        at com.aerospike.client.async.NioEventLoop.runCommands(NioEventLoop.java:223)
        at com.aerospike.client.async.NioEventLoop.run(NioEventLoop.java:181)
        at java.base/java.lang.Thread.run(Thread.java:1583)

  TestAsyncTxn.asyncTxnWriteRead:123->execute:341->TestAsync.waitTillComplete:210 java.lang.Exception: Data mismatch: Expected val1. Received val2
        at com.aerospike.test.async.TestAsync.assertBinEqual(TestAsync.java:78)
        at com.aerospike.test.async.TestAsyncTxn$GetExpect$1.onSuccess(TestAsyncTxn.java:483)
        at com.aerospike.client.async.AsyncRead.onSuccess(AsyncRead.java:80)
        at com.aerospike.client.async.NioCommand.finish(NioCommand.java:822)
        at com.aerospike.client.async.NioCommand.readSingleBody(NioCommand.java:566)
        at com.aerospike.client.async.NioCommand.readSingleHeader(NioCommand.java:554)
        at com.aerospike.client.async.NioCommand.read(NioCommand.java:491)
        at com.aerospike.client.async.NioCommand.processEvent(NioCommand.java:325)
        at com.aerospike.client.async.NioEventLoop.runCommands(NioEventLoop.java:223)
        at com.aerospike.client.async.NioEventLoop.run(NioEventLoop.java:181)
…