@BrianNichols 2024-11-26
The referenced bug report indicates it was fixed in java 17.  Therefore, we can fix in aerospike-client-jdk21 (master branch), but not in aerospike-client-jdk8 (jdk8 branch).

Those Txn tests require server 8.0, so they are expected to fail on older server versions.