@BrianNichols 2024-12-03
Java client 9.0.1 has been released: https://download.aerospike.com/download/client/java/notes.html#9.0.1